feat(TFD-175): Refactor packages
