+++ components/components-processing/processing-definition/src/test/java/org/talend/components/processing/definition/replicate/ReplicateDefinitionTest.java
@@ -30,6 +30,7 @@ import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
 public class ReplicateDefinitionTest {
+
     private final ReplicateDefinition definition = new ReplicateDefinition();
 
     /**
