+++ components/components-processing/processing-runtime/src/main/java/org/talend/components/processing/runtime/replicate/ReplicateRuntime.java
@@ -1,4 +1,4 @@
-package org.talend.components.processing.runtime;
+package org.talend.components.processing.runtime.replicate;
 
 import org.apache.beam.sdk.values.PCollection;
